+++ src/ui.h
@@ -17,7 +17,7 @@
 #define X true
 #define Y false
 
-#define BARS_SZ 3
+#define BARS_SZ 2
 
 typedef int16_t buff_t;
 typedef uint16_t term_t; // TODO: MERGE THESE TYPEDEF WITH LOGIC_H.
@@ -26,9 +26,8 @@ void cursor(void);
 buff_t decIntLen(buff_t chars);
 term_t getSize(bool axis);
 void printDynamicString(const char *string, term_t max_len);
-void upperBar(const char *fname);
-term_t autoFill(struct Params buff, char key, term_t fill);
-void lowerBar(struct Params buff, char key);
+void bar(struct Data buff, char key);
+term_t autoFill(struct Data buff, char key, term_t fill);
 void cleanFrame(void);
 
 #endif