+++ docs/CHANGELOG.md
@@ -7,7 +7,7 @@ Changelog contains only new tangible features. To see what's new in source
 files use git-show or git-log.
 
 ## Versions:
-### v1.0.0 (WIP):
+### v1.0.0 (July 15, 2018):
 - Simplified API.
 - Less memory usage.
 - Proper text allocation.